+++ scanpipe/pipes/input.py
@@ -77,6 +77,14 @@ def get_tool_name_from_scan_headers(scan_data):
         return tool_name
 
 
+def get_extra_data_from_scan_headers(scan_data):
+    """Return the ``extra_data`` of the first header in the provided ``scan_data``."""
+    if headers := scan_data.get("headers", []):
+        first_header = headers[0]
+        if extra_data := first_header.get("extra_data"):
+            return extra_data
+
+
 def is_archive(location):
     """Return True if the file at ``location`` is an archive."""
     return get_type(location).is_archive
@@ -95,10 +103,13 @@ def load_inventory_from_toolkit_scan(project, input_location):
     )
 
 
-def load_inventory_from_scanpipe(project, scan_data):
+def load_inventory_from_scanpipe(project, scan_data, extra_data_prefix=None):
     """
     Create packages, dependencies, resources, and relations loaded from a ScanCode.io
     JSON output provided as ``scan_data``.
+
+    An ``extra_data_prefix`` can be provided in case multiple input files are loaded
+    into the same project. The prefix is usually the filename of the input.
     """
     for package_data in scan_data.get("packages", []):
         pipes.update_or_create_package(project, package_data)
@@ -111,6 +122,11 @@ def load_inventory_from_scanpipe(project, scan_data):
 
     for relation_data in scan_data.get("relations", []):
         pipes.get_or_create_relation(project, relation_data)
+
+    if extra_data := get_extra_data_from_scan_headers(scan_data):
+        if extra_data_prefix:
+            extra_data = {extra_data_prefix: extra_data}
+        project.update_extra_data(extra_data)
 
 
 model_to_object_maker_func = {